feat(api): Add reports controller and moved all reports endpoints to it
